requirements.txt
ollama
PyYAML
tqdm
Pillow